Dashboard - Fix up diagnostics
This will fetch all existing diagnostics for a given test path, and fix
up any issues with their start/end revisions (as well as delete any
duplicates). If tests are uploaded out of order, or a test is migrated
onto another, the diagnostics can end up completely out of order.

Bug: catapult:#4094
Change-Id: Iec2a0180027f1b104f4c388c89bbc052a9a0e527