Dashboard - InstallHooks() in testing_common setup.
This is causing issues when run with -j1 since they're in-process and
the stubs get overwritten by the testbed teardown.

Also provide a default value for sherrif.internal_only, as this
exposed that these are getting set to None in tests. It's now defaulted
to False, which is the same as Masters/Bots/Rows.

Bug: catapult:#4173
Change-Id: I082d7bc7d31e7981a09e66afa2210723b6adfd74