[Dashboard] Fix logic for add a_tracing_uri to histogram Rows
We have been checking whether there is exactly one trace URL per histogram,
which is always false for benchmarks that repeat stories multiple times.

The way chart JSON currently deals with this is only uploading the trace from
the last story run, which is a known issue that should be fixed comprehensively
by V2SPA's pivot table section. In the meantime, we emulate this behavior by
taking the last of the TRACE_URLS diagnostic.

Bug: catapult:#4414
Change-Id: I0549cfb029e83057c23f729c6188405f53801cb2