Update patch set 1
Patch Set 1:

I'll defer to Simon here.

My rationale for having it be a decorator:
Authentication is really easy to get wrong on the web. It's very easy to have an entry point to a function that checks auth and has an accidental fallthrough. If we want all the authentication for the API to be the same, having decorators simplifies the process greatly.

Patch-set: 1
Reviewer: Gerrit User 1001913 <1001913@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>