Dashboard - Add a path for inserting out-of-order diagnostics.
There's 2 general cases, either you're submitting the latest revision
or an older one.

Nearly all cases will be the latest, so we just follow the previous
code path to do that.

Older or out-of-order commits need to be handled with a slower path. In
that, there's 2 general cases, you're clobbering an older diagnostic or
splitting an existing range. We assume when an out-of-order commit is
added with a differing diagnostic, that the diagnostic is valid up
until the next commit after that point.

ie. If you had commits at [0, 10, 20, 50, 100], and one diagnostic that
spanned the entire range [0, sys.maxint, 'foo1']

If you were to split the range, you'd find the next commit since at
each commit the previous diagnostic was diffed during upload and found
to be the same. So inserting a new 'foo2' at revision 25, the range
would look like this:

[0, 24, 'foo1']
[25, 49, 'foo2']
[50, sys.maxint, 'foo1']

Bug: catapult:#4094
Change-Id: I7190309ac0bb73c1cd2d6990b7f2eaa38ee55526