Update patch set 2
Patch Set 2:

This looks great. Hadn't known about your "glop" parser generator before now, but seems like a good idea to me to use it, given that the grammar for the test expectation files is non-trivial. Will we be able to bring up a browser instance in https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/testing/run_browser_tests.py so that we can pass the right --expectations-tag arguments for the current platform?

How can we make more progress on this? Should I patch this into my workspace and try to get it working with SeriallyExecutedBrowserTestCase and the GPU tests? It seems non-destructive to typ, so seems OK to me to land and then we can iterate on it, but also happy to help make sure it works for one of our test types first.

Patch-set: 2
Reviewer: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>