[pinpoint] Add isolatedserver field to Swarming tasks.
Swarming servers have a default Isolate server if you don't specify
it. This change allows us to send a Job to a Swarming server with a
different Isolate server.

During the transition, we will temporarily use the internal Swarming
server with the external Isolate server, so we need to tell the
Swarming server which Isolate server to use.

We're configuring the builders to upload the isolate server with the
isolate hash, so we can store the isolate server in the Datastore for
each isolate and look it up from there when running a Job.

Bug: catapult:#822008
Change-Id: I33d7176ad364843007add86352d9f89ed21ff34a