[soundwave] Add pandas_sqlite.InsertOrReplaceRecords
Pandas DataFrame.to_sql method has limitation of not being able to
"insert or replace" records, see e.g:
https://github.com/pandas-dev/pandas/issues/14553

Using pandas.io.sql primitives, however, it's not too hard to implement
such a functionality (for the SQLite case only).

Assuming that the dataframe has a simple named index (single column),
this method will use that column for the PRIMARY KEY of the table.

Bug: catapult:#4382
Change-Id: I5f70a04c18b998590b1e77e5ff6b89b2d27138af