Update patch set 6
Patch Set 6:

The CL looks great to me overall. Except I would try to avoid adding APIs in browser.py and only expose those APIs in android_browser_backend at most. It's fine to access _foo within catapult/Telemetry.


The way to reason about this is the X class in Telemetry are generally public and can be accessed by users who write benchmarks/pages in chromium/src/tools/perf. For that reason, we typically try to keep X's Api minimal so that future API changes is easier (example: a simple API update like changing a parameter to not having default value in Telemetry public API takes ~50 CLs in https://bugs.chromium.org/p/chromium/issues/detail?id=720002)

But how do we make sure that it's still easy to develop new features like this patch? 
As an escape hatch, we have X & X_backend construct in Telemetry. X is the minimal API version which is public, whereas X_backend is the more complex set of APIs that are only exposed to Telemetry's internal code. One can think about this as C++ friend class design pattern, or Java's protected package access feature (APis that are only visible to a limited number of places)

I defer the rest of the review to Charlie. Thanks again for your patience with us working on this!

Patch-set: 6