Telemetry: Add benchmark info to shared diagnostic
The CL fixes these issues:

1- It is added to individual histograms right now.
2- If there is not --results-label specified, run_benchmark
   will create a separate column for benchmark duration
   because we are setting the label of that histogram to
   None.

Also, this is in preparation for my next CL that will
attempt to add both legacy and TBMv2 metrics to results (for
the rendering benchmark). We want TBMv2 histograms don't
miss benchmark info and so have the same label as legacy
histograms from the same run.

Change-Id: I76b52eae88c38b19257c2927e73ade845b060c9a