Add ActionRunner.ProfilingInterval
When running locally for testing, a test procedure can be modified to
generate a CPU profile over particular test actions of interest by
enclosing the actions in a ProfilingInterval context; for example:

def TestProcedure(self, action_runner):
  action_runner.Navigate(self.test_url)
  with action_runner.ProfilingInterval("profile.pb"):
    action_runner.RepaintContinuously(10)

Note that this is only known to work on Linux and Mac, as the
underlying code in chrome relies on tcmalloc.  Additionally, the
browser must be given the '--no-sandbox' flag to permit the renderer
process to write to the file system.

Change-Id: I79ba663eef3273a05d9e8f6b988e52c4d8101142