Refactor story_runner to use shared_state's platform for determining test disabling state
This helps achieve two things:
1) It makes sure that Telemetry benchmark_runner don't invoke browser discovering two times:
one to discover the target test platform for test disabling logic, and another one
to actually run the story run loop.
2) It makes the implementation of story_runner module less specific on the browser logic
since the logic of discovering the platform which the test to run on is now delegated to
shared_state (which is customizable by clients). This helps fixing the issue where the
"browserless" cluster telemetry benchmark breaks on machine that don't have browser (see
attached bug).

Bug: chromium:840428
Change-Id: If18dbbe2684c279f55b323c7523229febb0ee19c