[dashboard] Add Google Surveys to alerts, report, group_report, speed_releasing.
This CL adds the Google Surveys script to the /alerts, /report, /group_report,
and /speed_releasing pages on the chromeperf dashboard.
For now, all 4 pages are treated as the same site to avoid bombarding the user
with too many surveys. If necessary, a followup CL could change the site IDs in
these script tags to separate site IDs.
This CL does not manually trigger surveys. Google Surveys decides when to
display them, though this could also be changed in a followup CL.
A single satisfaction survey is currently running for this site as a pilot to
prove out the process. It will close after collecting 50 responses, then we can
evaluate and change things and launch new surveys from surveys.google.com
without changing these script tags. Future surveys for the current dashboard
will be used to obtain a baseline satisfaction score for comparison against V2SPA.
Adblock blocks these surveys.

https://imgur.com/txyw329.png
Change-Id: I4c120b314139ea09be32079d51f39f51d8bf46c8