Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I don't fully understand how this work, so hesitate to allow adding this to core Telemetry API. 
> > 
> > I would suggest adding this as a util method "def ProfileInterval(action_runner)" in tools/perf/, probably next to repaint_helper.py
> 
> Now that you have some more background, what do you think about this patch? I see three ways to approach this:
> 
> 1) Land the patch as-is, with the intention that devs will patch the benchmark test code.
> 
> 2) Land the patch in src/tools/perf, with the intention that devs will patch the benchmark test code.
> 
> 3) Land the patch more or less as-is, and make it behave the same as the simpleperf options, i.e, when running on linux, it will generate a profile over the "navigation" and/or "interactions" part of page tests.
> 
> I think it would be nice if the --simpleperf-* arguments could be made generic, and generate profiles during the "navigation" and/or "interactions" phases on either android or linux. But again, I really don't know what your intention is for the existing --profiler options, and I don't want to keep piling on redundant features.

My intention for the existing --profiler options is to eventually replace it with perfetto system. 

From here to there, we should try to minimum the API surface we expose to devs, so that the work of deprecating the old & replacing with the new.

With this in mind, I think it make most sense to integrate this to the existing BrowserSimpleperfController. My suggestion:

1) Rename BrowserSimpleperfController to BrowserSamplingController
2) Modify BrowserSamplingController so that it do simpleperf on Android, and invoke chrome.gpuBenchmarking.startProfiling on linux desktop


Unrelated, how commandline flag are we using to activate simpleperf now? We should make sure Telemetry add --no-sandbox to the browser startup argument if user enable sampling flag on Linux desktop

Patch-set: 4
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>