Update patch set 4
Patch Set 4:

> My intention for the existing --profiler options is to eventually replace it with perfetto system. 
> 
> From here to there, we should try to minimum the API surface we expose to devs, so that the work of deprecating the old & replacing with the new.

While perfetto promises to be a great addition, we should keep in mind that it's still in pretty early stages, so won't be available for a while. CPU profiling is not included in the initial feature set, so it's very unclear when it will be available. This is based on conversation with Primiano at BlinkOn. In my opinion, at this point it doesn't make sense to reserve a parking space in the telemetry code for perfetto.


> With this in mind, I think it make most sense to integrate this to the existing BrowserSimpleperfController. My suggestion:
> 
> 1) Rename BrowserSimpleperfController to BrowserSamplingController
> 2) Modify BrowserSamplingController so that it do simpleperf on Android, and invoke chrome.gpuBenchmarking.startProfiling on linux desktop

I agree that we'll need a name that is not specific to simpleperf, but "sampling" is really not descriptive of this functionality. The most accurate term is "profiling". If we need something to distinguish the new profiling code from the old, then perhaps "interval profiling" could be the term d'art.

I still don't really understand the current status of the profiling code. You announced deprecation over a year ago. Is there any reason not to just rip the old code out? It seems like your original intention was to merge the profiling functionality into the tracing code. However, it should be clear now that profiling does not fit well into the tracing pattern (or at least that it would require invasive changes to the tracing code).


> Unrelated, how commandline flag are we using to activate simpleperf now? We should make sure Telemetry add --no-sandbox to the browser startup argument if user enable sampling flag on Linux desktop

--no-sandbox is not required for simpleperf, it is specific to chrome.gpuBenchmarking.startProfiling. So yes, if that functionality is exposed on the command line, then the telemetry code would need to add --no-sandbox to the browser command line.

Patch-set: 4
Reviewer: Gerrit User 1001706 <1001706@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>