Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> (1 comment)
> 
> Why do we need both profile_files_to_copy and profile_data_to_copy ? For the profile_data_to_copy, can the callsite just write the data to some tempfile & reuse profile_files_to_copy flag?

I tried that using python tempfile module, but it seemed tricky.

E.g. the most "natural" implementation would be to use tempfile.NamedTemporaryFile, and scope that file to the lifetime of the PerfBenchmark in tools/perf/core/perf_benchmark.py.

However, the docs say:
"""
Whether the name can be used to open the file a second time, while the named temporary file is still open, varies across platforms
"""

This means we probably shouldn't be using NamedTemporaryFile. The alternative I thought of was using some tempfile method of writing a file (and closing it), then deleting it when PerfBenchmark is destroyed in a __del__ method. I'm not a python expert but if that solution seems good to you I could try to implement it.

There are a few other options I thought of:
1. Sending the raw data to telemetry, with a profile path to write it to
2. Sending the raw file objects to telemetry

Happy to hear your thoughts!

Patch-set: 4
Reviewer: Gerrit User 1142059 <1142059@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>