Dashboard - Remove UI features that use "monitored"
There are 2 usages on the UI, at the suite level to display that little
hint in the test-picker that says "monitored" or "unmonitored".

The 2nd is in chart-container, to determine if a test is important but
doesn't seemingly make much difference (possibly because of Ethan's
changes to list_tests last year).

Demo: https://dev-simonhatch-d6d0bcb3-dot-chromeperf.appspot.com/report

Bug: catapult:#3930
Change-Id: I8a17f1963b63895ebec8c1c005fe4e2b869eb06d