Update patch set 2
Patch Set 2:

Not sure if we should store `comparison_mode` in both the Job and JobState, or if JobState should expose it as a property so that Job can find it from JobState. The advantage of storing it in the Job is that it's indexable and queryable. You would be able to filter the job list for perf try jobs vs. bisect jobs.

Patch-set: 2