Introduce a BrowserOption for seeding a default profile
This will be used for ad-tagging work, which needs two pieces:
1. Copying a generated file into <user-data-dir>. This file is a ruleset
   which Chrome uses to compute whether resources match ad requests.

2. Seeding some JSON prefs. We need to have the initial Local State prefs
   seeded with the correct ruleset version, otherwise it won't check the
   for the file that we copied in (1).

Both (1) and (2) can be satisfied using an API that can copy over
exiting files to the profile, when it is created. This CL adds a new
BrowserOption, |profile_files_to_copy|, which takes a list of tuples
(source_path, dest_path), where dest_path is relative to the created
profile, and copies files in source_path to dest_path on creation.

This CL adds support for this option in PossibleDesktopBrowser. Follow-ups
will add Android support. See crrev.com/c/1054590 for the Chromium CL
which uses this new API.

This CL also fixes a small bug when setting browser_options.profile_dir
in PossibleDesktopBrowser.

Bug: chromium:823982
Change-Id: I3ea8c4f3231fa145c0cf278afc2d16995b73746d