Refactor Histogram relationship diagnostics.
Demos:
http://www/~benjhayden/3009553002.html
http://www/~benjhayden/3009553002-loading.html

SPLIT
- add Breakdowns to loadingMetric
- add Breakdowns to <what other metrics?>
- translate existing RelatedHistogramMap/Breakdowns to RelatedNameMaps
- remove RelatedHistogramMap/Breakdowns
- refactor HistogramSet to map from name to set of histograms

Currently, Histogram relationships are encoded as RelatedHistogramMap and
RelatedHistogramBreakdown diagnostics.
RelatedHistogramBreakdown is a histogram-level analog to the sample-level
Breakdown diagnostic, but there is no encoded relationship between them.
RelatedHistogramBreakdown is a subclass of RelatedHistogramMap.
RelatedHistogramMap uses UUIDs to refer to related Histograms directly.

Merging Histogram relationship diagnostics is very complex.

There are not actually any uses that require Histograms to be related
directly (via guid). All known uses of Histogram relationships actually care only
about the names of related Histograms.
Metric authors think of relationships as being between histogram names, not between
histogram objects.

The "relatedNames" diagnostic name is reserved for a GenericSet synthesized from
RelatedHistogramMaps and RelatedHistogramBreakdowns.

breakdown-span visualizes both Breakdowns and RelatedHistogramBreakdowns. The only
difference between them is that RelatedHistogramBreakdowns are visualized as links
which, when clicked, send the names of the related Histograms up to the
histogram-set-table in an event so the table can filter out unrelated rows.
Breakdown-span cannot linkify Breakdowns.

The distinction between RHBs and Breakdowns is an architectural wart.
There should be a single way for metrics to encode sample breakdowns, and its
visualization should be linkable.

This CL replaces RelatedHistogramMap and RelatedHistogramBreakdown with
RelatedNameMap, which is a Map from string key to a Histogram name.
RelatedNameMaps could be visualized as a set of analysis-links, but don't need to
be since breakdown-span produces those analysis-links by correlating Breakdown
entries with RelatedNameMaps.

When metrics create RelatedNameMaps directly, there is no need for "relatedNames"
or addRelatedNames().

Merging RelatedNameMaps is 2 orders of magnitude simpler than
RelatedHistogramMap/Breakdown.

https://github.com/catapult-project/catapult/issues/3761

Change-Id: I7f9d1116ceda08b5acaf0df4e59564cf723957ee