[pinpoint] Store `comparison_mode` in JobState.
JobState doesn't have access to `comparison_mode` on the
Job. But, there are a few places where we want to use
`comparison_mode` inside JobState. For example, we calculate
the result_values differently for functional and performance
jobs. We might also use different significance thresholds or
repeat counts for each type. (crbug.com/842163)

Bug: catapult:#3871
Change-Id: I156755510a520501615f365ae9786bb0ceeef3f3