Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 2:
> > 
> > Not sure if we should store `comparison_mode` in both the Job and JobState, or if JobState should expose it as a property so that Job can find it from JobState. The advantage of storing it in the Job is that it's indexable and queryable. You would be able to filter the job list for perf try jobs vs. bisect jobs.
> 
> I'm not opposed no neither this CL (moving comparison_mode to JobState only); or even have comprarison_mode in both JobState and Job (e.g. to be able to index by comparison_mode, that sounds like a good reason).
> 
> The only thing that appears slightly strange is that JobState is not able to access information stored in its corresponding Job. Would it be possible to solve that issue in a more general way?

I'm going to keep this CL as-is, with no comparison_mode in Job, since it's simpler and we can add it to Job later if we need the feature.

I think it's fine if JobState can't access Job. I think of the data as flowing only in one direction -- Job gathers all of the necessary information, then sends JobState off to do its thing. Job deals with interacting with users and external services, and JobState is the workhorse, running the actual Job.

Patch-set: 4
Reviewer: Gerrit User 1000093 <1000093@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>