[trace-viewer] Fix "no tree" error in layer quad stack view
The error occurred when there were input events in the trace, and
"input events" was enabled in the layer quad stack view. It was because
the image of input event was never loaded so the view was never updated.
The failure was because the image loading was started before the view
was added in the DOM tree when the image didn't have computed style yet.

Bug: catapult:#4471
Bug: chromium:842238
Change-Id: I705bf766c9e8d2b91991c3f4261dfaabeda7aab2