Change "minutes" legacy unit to timeInMsAutoFormat.
The 'minutes' unit is used by telemetry to record benchmark duration.
When converted to Histograms, minutes are converted to ms, which results in
unreadably long numbers.
Changing this unit to timeInMsAutoFormat will allow V2SPA and results.html to
render benchmark durations as minutes or hours, much easier to read.

Bug: catapult:#4461
Change-Id: I6c157519dffb7d58dcd09f33f5e3f617f7aae6bd