Use ResizeObserver for picture chart resizing
ResizeObserver has been available since Chrome 64.
It provides more complete solution to respond to resize.

The previous code handled resize of the charts by the chagne
of attributes of the left panel only, which missed a lot of
other cases (e.g. the right panel resized because of window
resize).

Bug: catapult:#4471
Bug: chromium:842238
Change-Id: I0187ca86f2aad3a0b09139c91b236a27c3fdc512