[pinpoint] Generate Execution details URLs on the server side.
To launch Pinpoint, we hacked in hard-coded URLs into the UI for the
Swarming, isolate, and build links. Now that we have jobs running on
the internal Swarming server, those links are wrong. If we instead
generate these URLs on the server, we have the additional context
needed to produce the correct URLs.

Bug: catapult:#3899, catapult:#4488
Change-Id: Ie656da12cb8e86382f00010647c1d5712962b701