Add support for profile_files_to_copy in PossibleAndroidBrowser
This CL also removes AndroidPlatformBackend.RemoveProfile, which is
now unused. For clients with empty profile_files_to_copy, this
method is replaced with a call to PushProfile with an empty source
directory.

This is intended to have no behavior changes, and the removal of
AndroidBrowserBackendSettings.profile_ignore_list seems to be
OK given how _EfficientDeviceDirectoryCopy is implemented. See
|delete_extra| in efficient_android_directory_copy.sh, which mentions
protecting |lib|.

Bug: chromium:823982
Change-Id: I279bf625fb233ad263d500605d940ace0639751a