Add some small indexes over Anomaly.
Currently, there are a few 3-, 4-, and 5-property indexes over Anomaly to support
a few specific queries.
V2SPA and the API will be using several new queries over Anomaly.
It would be expensive to create a separate index for every possible combination of
properties.
This CL creates a few new indexes over pairs of properties in order to allow the
datastore to use the zigzag merge join algorithm to support any combination of
properties.
https://cloud.google.com/appengine/articles/indexselection

https://dev-benjhayden-1cad5f3b-dot-chromeperf.appspot.com/_ah/dev_console/interactive

Bug: catapult:#4479
Change-Id: Iea0191eaedcd52b5e87e820a2789f72bb2ce3be5