Replace the Anomaly indexes with several 2-property indexes.
Currently, there are a few 3-, 4-, and 5-property indexes over Anomaly to support
a few specific queries.
V2SPA and the API will be using several new queries over Anomaly.
It would be expensive to create a separate index for every possible combination of
properties.
A single index over all properties would only support a single query that
specified all properties, not any query that only specified a subset of
properties.
This CL creates a few new indexes over pairs of properties in order to allow the
datastore to use the zigzag merge join algorithm to support any combination of
properties. The existing 3-, 4-, and 5-property indexes can be removed without
changing the code.
https://cloud.google.com/appengine/articles/indexselection

Bug: catapult:#4479
Change-Id: Iea0191eaedcd52b5e87e820a2789f72bb2ce3be5