Update patch set 2
Patch Set 2:

Oops, sorry for the latency here and feel free to ping in the future, this somehow slipped off my plate. Also, with the caveat that I'm not the most familiar with this code, I'm going to try my best to review.

I'm confused about this, because we already call this method here https://cs.chromium.org/chromium/src/third_party/catapult/tracing/tracing/importer/import.html?l=213&rcl=517100ca073f7deaf4818786104daf7e7c744b8d

Why is that not enough? Maybe the core bug here (and I skimmed the bug, please correct me if my understanding is flawed) is that ftrace_importer wrongly makes assumptions about events that it has no jurisdiction over. I wonder if that's something we should mitigate instead. As-is, this fix feels a little band-aidy to me, because it sounds like we're saying that we should change the behavior of import due to an assumption that ftrace_importer wants to make.

Patch-set: 2
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>