Generalize the /api/alerts request handler.
The alerts api request handler currently supports a few specific queries,
and only supports HTTP POST.
Soundwave and V2SPA require more flexibility.
This CL adds a GET handler that accepts 0-17 query parameters to allow
callers to construct almost any query. Callers may optionally pass an
Authorization header; if not supplied, only non-internal_only alerts are
returned.

Queries that work with the current indices:
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?sheriff=Chromium+Perf+Sheriff
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&cursor=CkIKFgoJdGltZXN0YW1wEgkIhKCjg8Ow2wISJGoMc35jaHJvbWVwZXJmchQLEgdBbm9tYWx5GICAkLHN4aIKDBgAIAE=

Queries that do not work with the current indices, but would with an Authorization header:
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?bug_id=846845

Queries that do not work with the current indices, even with an Authorization header:
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&recovered=false
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&test_suite_name=system_health.memory_mobile&is_improvement=true&bot_name=health-plan-clankium-phone
https://dev-benjhayden-a9015a14-dot-chromeperf.appspot.com/api/alerts?limit=10&test_suite_name=system_health.memory_mobile&is_improvement=true&bot_name=health-plan-clankium-phone

Another CL restructures the Anomaly indexes in order to support the
generality of this new handler:
https://chromium-review.googlesource.com/c/catapult/+/1077070

Bug: catapult:#4479
Change-Id: I92c2770a89fe1769503afdabe2e05ba1c96e0549