Remove supports_system_info.
Remove supports_system_info property, and updated the callers to
directly call GetSystemInfo() and check for its existence instead.
supports_system_info makes a call to GetSystemInfo(), which triggers a
request back to the browser, and immediately discards the response. So,
instead of making two calls to the browser for GetSystemInfo(), make a
single call.

When running locally on a Z840, this saves ~1.5 seconds for each test.

BUG=chromium:0

Change-Id: Ie9d0a3d5ccd13f12722395c6e8d77feb4e1ae9a4