Fix ChangeSecurityContext command in DeviceUtils
chcon -R is not supported everywhere. This CL also adds a few simple
integration tests which exercise PushProfile / RemoveProfile to
verify the fix.

The CL also adds a new private API to RunShellCommand, which forces
the input command to be run with `su`. This apparently is needed to
avoid permission errors with `chcon` on some versions of Android.

Bug: catapult:#4498
Change-Id: I77de6623dc14d34a540c7f1204d6f1ff2e18d53d