Generalize the /api/alerts request handler.
The alerts api request handler currently supports a few specific queries.
Soundwave and V2SPA require more flexibility.
This CL adds support for a new URL pattern with 0-17 query parameters to allow
callers to construct almost any query. Callers may optionally pass an
Authorization header; if not supplied, only non-internal_only alerts are
returned.

Queries that work with the current indices:
/api/alerts
/api/alerts?bug_id=846845
/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10
/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&cursor=CkIKFgoJdGltZXN0YW1wEgkIhKCjg8Ow2wISJGoMc35jaHJvbWVwZXJmchQLEgdBbm9tYWx5GICAkLHN4aIKDBgAIAE=

Queries that do not work with the current indices:
/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&recovered=false
/api/alerts?sheriff=Chromium+Perf+Sheriff&limit=10&test_suite_name=system_health.memory_mobile&is_improvement=true&bot_name=health-plan-clankium-phone
/api/alerts?limit=10&test_suite_name=system_health.memory_mobile&is_improvement=true&bot_name=health-plan-clankium-phone

Another CL restructures the Anomaly indexes in order to support the
generality of this new handler:
https://chromium-review.googlesource.com/c/catapult/+/1077070

Bug: catapult:#4479
Change-Id: I92c2770a89fe1769503afdabe2e05ba1c96e0549