Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > lgtm, but probably also wait for Simon.
> > 
> > Maybe the "Analyze benchmark results" link should be renamed and turned into a link to download the results2.html file?
> 
> I'm not sure what we should do for that yet. Do users ever use results2 for bisect jobs?

I might use them. :)

> This change applies only to try jobs with no comparison_mode.

And, yes, I was thinking about the case with no comparison mode; in the demo both the iframe with results2 *and* the "Analyze benchmark results" link  to open results2 are shown. I think the later is not as useful if I'm already looking at the iframe; but a link to download the file might be handy.

Patch-set: 2