Update patch set 16
Patch Set 16:

I assume this has to be rebased on the previous CL, so deferring full review after that lands.

Some high level comments, though: I understand what this CL does is (basically) ensuring that the base dirname for the profile on the host has a stable known value ("_default_profile"), so it becomes easier to avoid littering the sdcard. Right?

To me that sounds that this should be an implementation detail of PushProfile (and not of SetupEnvironment). Also, wondering if there is another solution that avoids creating a second copy on the host? (not sure, just wondering now ...)

Patch-set: 16
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>