Update patch set 17
Patch Set 17:

(2 comments)

> Patch Set 16:
> 
> I assume this has to be rebased on the previous CL, so deferring full review after that lands.
Yep, this has a dependency on the previous CL.

> 
> Some high level comments, though: I understand what this CL does is (basically) ensuring that the base dirname for the profile on the host has a stable known value ("_default_profile"), so it becomes easier to avoid littering the sdcard. Right?
Yes, that's true. Since now we'll push a profile almost every time with |profile_files_to_copy|, we don't want it to be random to avoid littering the sdcard.

> 
> To me that sounds that this should be an implementation detail of PushProfile (and not of SetupEnvironment). Also, wondering if there is another solution that avoids creating a second copy on the host? (not sure, just wondering now ...)

It's difficult to say. The behavior of PushProfile is to copy over exactly the input into /sdcard, and then copy files from /sdcard into the actual profile. This behavior can be tested in isolation. However, the input of the function is controlled by the caller and the caller needs to be aware of this behavior. PushProfile has no way of detecting if it is being used improperly.

There may be a solution without using the sdcard, but I'm not familiar enough with the performance analysis that went into the original solution. I guess just pushing files directly is too slow.

Patch-set: 17
Reviewer: Gerrit User 1142059 <1142059@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>