Update patch set 17
Patch Set 17:

> It's difficult to say. The behavior of PushProfile is to copy over exactly the input into /sdcard, and then copy files from /sdcard into the actual profile. This behavior can be tested in isolation. However, the input of the function is controlled by the caller and the caller needs to be aware of this behavior. PushProfile has no way of detecting if it is being used improperly.

I don't see it like that. I think the contract that PushProfile should be doing with its clients is just: Give me a path to a directory on the host, I'll make sure to populate the profile of the browser (determined by "package") using the files contained on that directory.

Whether we make extra copies on the host, or the device's sdcard, or other smart things to optimize multiple calls to this method with same/similar files, are just implementation details which the caller shouldn't need to worry too much about.

Patch-set: 17
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>