tabs: Ignore the duration for the new metric.
It is possible that in the new metric, the final results gets a mix
of the new metric and the old metric (based on slice-duration). There
can be un-terminated slices in some cases, and instead of using the
new metric for these, it ends up using duration, which is incorrect.
So explicitly look for new metric, and have some code to work for
the legacy cases too.

BUG=chromium:849182

Change-Id: Ia5982316eb0398f92e9b5c9cd58f864fc76d95fd