[soundwave] Ensure API and DB are always initialized
When executing either "alerts" or "timeseries" soundwave commands we
need to make sure that both the API credentials have been aquired and
DB tables have been initialized before sending off work to subprocess
pools.

To achieve, we factor out this initialization to a new method which
can also control the lifetime of the DB conneciton on the main
process.

Bug: catapult:#4503
Change-Id: I1d0da17551a9adad98f01a6e8dc5dd3cdb9dd0c2