First pass at supporting test expectations in TYP.
This patch adds support for test expectations to TYP, mostly by reusing
the expectation parser written for catapult. The expectations
themselves follow the spec in bit.ly/chromium-test-list-format.

Specify the paths to the expectations files in args.expectations_files
and the tags to use when running (e.g., ['Release', 'Mac', 'Mac10.13'])
in args.expectation_tags, and run as normal.

Bug: 807751

Change-Id: I28e4e5d9cde50c1fad5eecac88e940f02eb74815