[Sheriff] Disable telemetry.internal.browser.browser_unittest.BrowserTest.testForegroundTab
The test was failing uner MacOS bots. The test used to be enabled only
for 'has tabs'. We only need to disable it for 'mac', but there does
not seem to be way to mix @decorator.Enabled and @decorator.Disabled,
so it's safer to just disable the test always.

NOTRY=true
TBR=aberent

Bug: 851463
Change-Id: I8d555eadf33087ecfafbf07faa37389d53fc8d2e