Telemetry: Scroll up at the end of test PageTests
The browser is left open after each run on CrOS:

https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/page/shared_page_state.py?l=177

So, for example, when pageset_repeat is more than 1, at the 2nd and
later runs, the page is already scrolled down, which means 2nd and
later runs don't generate animations. This can cause tests that set
pagetest_repeat to more than 1 see unexpected results.

One way to fix, other than closing the browser at the end of each run,
is to make sure that the test PageTest undoes the scroll at the end.

Bug: chromium:851523
Change-Id: I086c78beab8f1306a52a31fd9de4bfaf70e147b6