[Build] HealthyDevices filter by ABI
This adds a filter DeviceUtils.HealthyDevices to limit to output
to a devices supporting a specified list of ABIs. This filter is
imposed by the abis kwarg and is an optional filter. If a filter
is imposed it will warn about other attached devices that don't
support the ABI in the APK. In order to do so the aapt module
needs to expose the "list" command to inspects the contents of
the lib/* folder. apk_helper.py uses aapt to extract the ABIs
supported by the APK.

Bug: 853280
Change-Id: Ic36cb18169c31dc3d0ec2b7ff9cb37455445e19b