[Telemetry] Example: restart browser during story
A few bugs prevent us from restarting the browser for the duration
of a story (crbug.com/854212), this CL includes an example of
workarounds that a benchmark can use to avoid those bugs.

Also includes a small change in tab.Close() to allow closing the
last tab of the browser. This was previously prevented by the tab
backend. By default the old behavior is retained, but an option to
override this is now provided.

Bug: chromium:849703
Change-Id: I44000226a69100366c1cc04a1378badd219421f6