[Telemetry] Add clear_caches option to possible_browser.Create
This is a first step towards moving browser cache-clearing decisions
out of the possible-browser and into the shared state.

At the moment a clear_caches option is added, which defaults to True
and maintains the current behavior.

Note: we do change, however, the behavior of Android's
FindExistingBrowser, for which we switch off cache-clearing already.

On follow up CLs, shared state implementations can individually
take control over cache-clearing responsibilities and switch this
flag to False to prevent "double-clearing".

Finally, when all shared states have been migrated, both the
option and the cache-clearing code in possible-browser can be
removed.

Bug: chromium:811244
Change-Id: Ie0c2fb0c5f1665e2bf0383943b182bd6afc30135