Add 90th percentile to results.html.
Currently, results.html allows users to choose which statistic to display in
closed cells. The options that it offers are collected from the statistics
configured by metrics via customizeSummaryOptions. Results.html ensures that it
always offers 'avg' even if no metrics configure that statistic.
However, it is common for histograms to have outliers, and for users to want to
ignore those outliers somehow.
This CL ensures that results.html also always allows users to display the 90th
percentile in closed cells even if no metrics configure that statistic.
This is a small step towards a broader change in this statistic picker:
https://chromium-review.googlesource.com/c/catapult/+/731652

Change-Id: Idc0d85c9c5466797d1c083b36781027a97332fc1