[Telemetry] Fix and re-enable StartupTracingTest
Refactor code in this test in preparation to adding more tests for
other cases (e.g. closing the browser before tracing is stopped).

Now, instead of parsing the trace through TimelineModel, we run
a simple metric on it and check that no errors are produced.

This also re-enables the test on Android.

Bug: chromium:854212,catapult:#3099
Change-Id: I54e7b6ceb3b7eb54812ac4b93b20edf75c7660ec