Telemetry: frame times in TBMv2
This is the first CL in a series of CLs to write smoothness metrics
in TBMv2.

The new metric names end with _tbmv2, for now, but their value should
be exactly the same as the old metrics. So, as soon as we delete the
smoothness benchmark (crbug.com/855063), we can delete the python
definitions of these metrics and remove the _tbmv2 suffix from the
TBMv2 versions.

This CL introduces frame_times_tbmv2, mean_frame_time_tbmv2, and
percentage_smooth_tbmv2.

Bug: 760553
Change-Id: I5de5af6da0d9b11a99b888d1bfa2cb149532935e