Remove unused Row index.
See this comment for a summary of the Row queries:
https://github.com/catapult-project/catapult/issues/4440#issuecomment-394947024
The only query that orders by revision ascending is in
graph_data.GetRowsForTestBeforeAfterRev, and it does not project value, so it
does not use this index.

I'll clean this up tonight and keep an eye on the logs. If NeedIndexErrors or
timeouts show up, then I'll recreate it. It might take a few days to recreate,
but next week should be a light week for most users.

Bug: catapult:#4440
Change-Id: Icc623d6677c92c3d88317c3328ac7a2c4153c7b7