[Telemetry] Fix crash caused by unhandled atrace timeout
This change fixes a crash caused when the Systrace Atrace agent's
GetResults() method returns False if a timeout occurs.

With the existing impelentation a cryptic unhandled exception will be
thrown if a timeout occurs in GetResults():
"AttributeError: 'bool' object has no attribute 'raw_data'"

This change modifies the behavior when a timeout occurs to explitly
throw an exception to better inform the user of the timeout error.

Change-Id: I7ada753639ea26fc4f6ecf47f3685aaf83dc4ddd